+++ driver/src/main/java/oracle/nosql/driver/ops/Request.java
@@ -66,6 +66,11 @@ public abstract class Request {
     private RateLimiter writeRateLimiter;
     private int rateLimitDelayedMs;
 
+    /**
+     * @hidden
+     */
+    private boolean isRefresh;
+
     protected Request() {}
 
     /**
@@ -76,7 +81,11 @@ public abstract class Request {
         return timeoutMs;
     }
 
-    protected void setTimeoutInternal(int timeoutMs) {
+    /**
+     * @hidden
+     * this is public to allow access from Client during refresh
+     */
+    public void setTimeoutInternal(int timeoutMs) {
         if (timeoutMs <= 0) {
             throw new IllegalArgumentException("timeout must be > 0");
         }
@@ -415,6 +424,24 @@ public abstract class Request {
 
     public int getRateLimitDelayedMs() {
         return rateLimitDelayedMs;
+    }
+
+    /**
+     * @hidden
+     * internal use only
+     * @param value true or false
+     */
+    public void setIsRefresh(boolean value) {
+        isRefresh = value;
+    }
+
+    /**
+     * @hidden
+     * internal use only
+     * @return is refresh
+     */
+    public boolean getIsRefresh() {
+        return isRefresh;
     }
 
     /**